@@ -4,36 +4,44 @@ namespace BrandEmbassy\Doctrine\EnumType\Bridges\MarcMabeEnum;
 
 use BrandEmbassy\Doctrine\EnumType\EnumImplementation;
 use MabeEnum\Enum;
+use function assert;
+use function is_array;
+use function is_subclass_of;
 
-final class MarcMabeEnumBridge implements EnumImplementation
+/**
+ * @final
+ */
+class MarcMabeEnumBridge implements EnumImplementation
 {
-
     public function getBaseEnumClassName(): string
     {
         return Enum::class;
     }
 
+
     /**
-     * @param string $enumClassName
-     * @param string $databaseValue
+     * @param bool|float|int|string|null $databaseValue
+     *
      * @return mixed enum
      */
-    public function convertDatabaseValueToEnum(string $enumClassName, string $databaseValue)
+    public function convertDatabaseValueToEnum(string $enumClassName, $databaseValue)
     {
-        \assert(\is_subclass_of($enumClassName, Enum::class));
+        assert(is_subclass_of($enumClassName, Enum::class));
 
         return $enumClassName::get($databaseValue);
     }
 
+
     /**
      * @param mixed $enum
-     * @return string
      */
     public function convertEnumToDatabaseValue($enum): string
     {
-        \assert($enum instanceof Enum);
+        assert($enum instanceof Enum);
 
-        return (string)$enum->getValue();
+        $value = $enum->getValue();
+        assert(!is_array($value), 'Array is not supported as enum value when storing it to database');
+
+        return (string)$value;
     }
-
 }
