@@ -4,36 +4,42 @@ namespace BrandEmbassy\Doctrine\EnumType\Bridges\MyCLabsEnum;
 
 use BrandEmbassy\Doctrine\EnumType\EnumImplementation;
 use MyCLabs\Enum\Enum;
+use function assert;
+use function is_array;
+use function is_subclass_of;
 
-final class MyCLabsEnumBridge implements EnumImplementation
+/**
+ * @final
+ */
+class MyCLabsEnumBridge implements EnumImplementation
 {
-
     public function getBaseEnumClassName(): string
     {
         return Enum::class;
     }
 
+
     /**
-     * @param string $enumClassName
-     * @param string $databaseValue
      * @return mixed enum
      */
     public function convertDatabaseValueToEnum(string $enumClassName, string $databaseValue)
     {
-        \assert(\is_subclass_of($enumClassName, Enum::class));
+        assert(is_subclass_of($enumClassName, Enum::class));
 
         return new $enumClassName($databaseValue);
     }
 
+
     /**
      * @param mixed $enum
-     * @return string
      */
     public function convertEnumToDatabaseValue($enum): string
     {
-        \assert($enum instanceof Enum);
+        assert($enum instanceof Enum);
 
-        return $enum->getValue();
+        $value = $enum->getValue();
+        assert(!is_array($value), 'Array is not supported as enum value when storing it to database');
+
+        return (string)$value;
     }
-
 }
